+++ src/statscollector/DefaultStatsCollector.ts
@@ -186,7 +186,11 @@ export default class DefaultStatsCollector implements StatsCollector {
     if (clientMetricReport) {
       this.clientMetricReport = clientMetricReport;
     } else {
-      this.clientMetricReport = new DefaultClientMetricReport(this.logger);
+      this.clientMetricReport = new DefaultClientMetricReport(
+        this.logger,
+        this.videoStreamIndex,
+        this.audioVideoController.configuration.credentials.attendeeId
+      );
     }
 
     this.intervalScheduler = new IntervalScheduler(this.interval);
@@ -349,7 +353,8 @@ export default class DefaultStatsCollector implements StatsCollector {
 
   private getDirectionType(rawMetricReport: RawMetricReport): Direction {
     return rawMetricReport.id.toLowerCase().indexOf('send') !== -1 ||
-      rawMetricReport.id.toLowerCase().indexOf('outbound') !== -1
+      rawMetricReport.id.toLowerCase().indexOf('outbound') !== -1 ||
+      rawMetricReport.type === 'outbound-rtp'
       ? Direction.UPSTREAM
       : Direction.DOWNSTREAM;
   }